MathOptPresolve

MathOptPresolve.ModelStatus

MathOptPresolve.PresolveData

MathOptPresolve.PresolveTransformation

MathOptPresolve.ProblemData

MathOptPresolve.SolutionStatus

MathOptPresolve.bounds_consistency_checks!

MathOptPresolve.load_problem!

MathOptPresolve.postsolve!

MathOptPresolve.presolve!

MathOptPresolve.remove_empty_columns!

MathOptPresolve.remove_empty_rows!

MathOptPresolve.remove_fixed_variables!

MathOptPresolve.remove_forcing_rows!

MathOptPresolve.remove_free_column_singletons!

MathOptPresolve.@_return_if_inferred

MathOptPresolve.ModelStatus — Type

ModelStatus

NOT_INFERRED: Presolve has not inferred the status of the problem.

OPTIMAL: The current solution(s) are optimal.

PRIMAL_INFEASIBLE: The model is primal infeasible.

DUAL_INFEASIBLE: The model is dual infeasible.

MathOptPresolve.PresolveData — Type

PresolveData{T}

Stores information about an LP in the form

    min     c'x + c0
    s.t.    lr ⩽ Ax ⩽ ur
            lc ⩽  x ⩽ uc

whose dual writes

    max     lr'y⁺ - ur'y⁻ + lc's⁺ - uc's⁻
    s.t.     A'y⁺ -  A'y⁻ +    s⁺ -    s⁻ = c
               y⁺,     y⁻,     s⁺,     s⁻ ⩾ 0

MathOptPresolve.PresolveTransformation — Type

PresolveTransformation{T}

Abstract type for pre-solve transformations.

MathOptPresolve.ProblemData — Type

ProblemData{T}

Data structure for storing problem data in precision T.

The LP is represented in canonical form

\[\begin{array}{rl} \displaystyle \min_{x} \ \ \ & c^{T} x + c_{0} \\ s.t. \ \ \ & l_{r} \leq A x \leq u_{r} \\ & l_{c} \leq x \leq u_{c} \end{array}\]

MathOptPresolve.SolutionStatus — Type

SolutionStatus

Solution Status code

NO_SOLUTION: No current point.

FEASIBLE_POINT: The current point is feasible.

INFEASIBLE_POINT: The current point is infeasible.

INFEASIBILITY_CERTIFICATE: The current point is a proof of infeasibility

of the other problem. For example, a dual solution with INFEASIBILITY_CERTIFICATE status is a proof of infeasibility of the primal.

MathOptPresolve.bounds_consistency_checks! — Method

bounds_consistency_checks(ps)

Check that all primal & dual bounds are consistent.

TODO: If not, declare primal/dual infeasibility and extract ray.

MathOptPresolve.load_problem! — Method

load_problem!(pb, ...)

Load entire problem.

MathOptPresolve.postsolve! — Method

postsolve!

Perform post-solve.

MathOptPresolve.presolve! — Method

presolve(pb::ProblemData)

Perform pre-solve.

MathOptPresolve.remove_empty_columns! — Method

remove_empty_columns!(ps::PresolveData)

Remove all empty columns.

Called once at the beginning of the presolve procedure. If an empty column is created later, it is removed on the spot.

MathOptPresolve.remove_empty_rows! — Method

remove_empty_rows!(ps::PresolveData)

Remove all empty rows.

Called once at the beginning of the presolve procedure. If an empty row is created later, it is removed on the spot.

MathOptPresolve.remove_fixed_variables! — Method

remove_fixed_variables!(ps::PresolveData)

Remove all fixed variables.

MathOptPresolve.remove_forcing_rows! — Method

remove_forcing_rows!

Remove forcing and dominated row

MathOptPresolve.remove_free_column_singletons! — Method

remove_free_column_singletons!(ps)

MathOptPresolve.@_return_if_inferred — Macro

Helper macro for running a presolve routine and returning if that routine was able to infer the model status (e.g. optimal).