Primal reductions

MathOptPresolve.RemoveEmptyRow — Type

RemoveEmptyRow <: AbstractRule

Eliminate a single row if it is empty, i.e. all coefficients are zero.

In particular, consider row $i$ given as

\[l^{r}_{i} ≤ \sum_{j} a_{i, j} x_{j} ≤ u^{r}_{i}.\]

We eliminate row $i$ if $a_{i,j} = 0$ for each variable index $j$.

Presolve

If row $i$ is empty, the corresponding constraint is completely removed from the problem.

Infeasibility checks

For a prescribed tolerance $ϵ ≥ 0$, the problem is infeasible if either

row $i$ is empty and its lower bound is positive, i.e., $l^{r}_{i} > ϵ$, or

row $i$ is empty and its upper bound is negative, i.e., $u^{r}_{i} < -ϵ$.

The corresponding Farkas infeasibilities certificates are

$y^{l} = e_{i}$, $y^{u} = 0$, $z^{l} = z^{u} = 0$,

$y^{u} = e_{i}$, $y^{l} = 0$, $z^{l} = z^{u} = 0$,

where $e_{i}$ is the $i$-th vector of the canonical basis.

Postsolve

If the problem is feasible, $y^{l}_{i} = y^{u}_{i} = 0$.

If the problem is infeasible, see the Farkas rays above.

Misc

This is a primal reduction.

This reduction is non-destructive.

This reduction does not create any fill-in.

MathOptPresolve.RemoveEmptyRows — Type

RemoveEmptyRows <: AbstractRule

Remove all empty rows in the problem.

See RemoveEmptyRow.

MathOptPresolve.FixSingleVariable — Type

FixSingleVariable <: AbstractRule

Eliminate a single variable $x_{j}$ if its lower and upper bounds are equal.

Variable $x_{j}$ with lower and upper bounds $l_{j}, u_{j}$ is fixed to its lower bound if

\[| l_{j} - u_{j}| ≤ ϵ,\]

where $ϵ$ is a prescribed tolerance.

Presolve

Forall rows $i$,

\[l^{r}_{i} ≤ a_{i, j}x_{j} + \sum_{k \neq j} a_{i, k} x_{k} ≤ u^{r}_{i}\]

is transformed into

\[l^{r}_{i} - a_{i, j} l_{j} ≤ \sum_{k \neq j} a_{i, k} x_{k} ≤ u^{r}_{i} - a_{i, j} l_{j}.\]

The objective constant $c_{0}$ is updated to $c_{0} + c_{j} l_{j}$.

Infeasibility checks

The problem is infeasible if a variable is fixed to a value outside its domain.

Given an integer feasibility tolerance $ϵ_{int} ≥ 0$, infeasibility is declared in the following cases:

Variable $x_{j}$ is integer (or implied integer) and $l_{j}$ is fractional, i.e., $\min(l_{j} - ⌊ l_{j} ⌋, ⌈ l_{j} ⌉ - l_{j}) \geq ϵ_{int}$.

Variable $x_{j}$ is binary, and $l_{j} \notin [-ϵ_{int}, ϵ_{int}] \cup [1-ϵ_{int}, 1+ϵ_{int}]$.

Postsolve

For dual variables, first compute

\[z_{j} = c_{j} - \sum_{i} a_{i, j} (y_{i}^{l} - y_{i}^{u}),\]

then recover $z_{j}^{l} = z_{j}^{+}$ and $z_{j}^{u} = z_{j}^{-}$.

Misc

This is a primal reduction.

This reduction is non-destructive.

This reduction does not create any fill-in.