MathOptPresolve.apply! — Function

apply!(ps, rule, config) -> Nothing

Apply rule rule.

Arguments

ps::PresolveData{T}

rule::AbstractRule

config::PresolveOptions{T}

MathOptPresolve.bounds_consistency_checks! — Method

bounds_consistency_checks(ps)

Check that all primal & dual bounds are consistent.

TODO: If not, declare primal/dual infeasibility and extract ray.

MathOptPresolve.load_problem! — Method

load_problem!(pb, ...)

Load entire problem.

MathOptPresolve.postsolve! — Method

postsolve!

Perform post-solve.

MathOptPresolve.presolve! — Method

presolve(pb::ProblemData)

Perform pre-solve.

MathOptPresolve.remove_empty_columns! — Method

remove_empty_columns!(ps::PresolveData)

Remove all empty columns.

Called once at the beginning of the presolve procedure. If an empty column is created later, it is removed on the spot.

MathOptPresolve.remove_forcing_rows! — Method

remove_forcing_rows!

Remove forcing and dominated row

MathOptPresolve.remove_free_column_singletons! — Method

remove_free_column_singletons!(ps)