MathOptPresolve.apply! — Function

apply!(ps, rule, config) -> Nothing

Apply rule rule.

Arguments

ps::PresolveData{T}

rule::AbstractRule

config::PresolveOptions{T}

MathOptPresolve.bounds_consistency_checks! — Method

bounds_consistency_checks(ps)

Check that all primal & dual bounds are consistent.

TODO: If not, declare primal/dual infeasibility and extract ray.

MathOptPresolve.get_optimal_solution — Method

Query the optimal solution produced by presolve. The solution returned is for the original, not transformed, problem.

Note: Throws an error if presolve did not terminate with status MOI.OPTIMAL.

MathOptPresolve.get_unbounded_ray — Method

Query the unbounded ray produced by presolve. The ray returned is for the original, not transformed, problem.

Note: Throws an error if presolve did not terminate with status MOI.DUAL_INFEASIBLE.

MathOptPresolve.load_problem! — Method

load_problem!(pb, ...)

Load entire problem.

MathOptPresolve.post_crush — Method

post_crush(pr::PresolveResult{T}, x::Vector{T}; is_ray::Bool=false)

Map a point from the presolved model to the original model. If is_ray = false, the point should be a feasible point for the presolved model; otherwise, it should be an unbounded ray.

Note: Throws an error if the length of x is not equal to the number of variables in the presolved model.

MathOptPresolve.postsolve! — Method

postsolve!

Perform post-solve.

MathOptPresolve.presolve! — Method

presolve!(dest::MOI.ModelLike, src::MOI.ModelLike, T::Type{<:Real})::PresolveResult

Apply presolve to src model, and populate dest with the reduced problem. The type T specifies the data type used to represent the objective, constraints, etc. added to dest.

Returns a PresolveResult result object. You may query the status of the model after the presolve routine (get_status), and depending on this status, also query the relevant solution information (with get_optimal_solution, get_unbounded_ray, and get_infeasibility_certificate).

Note: The function will throw an ArgumentError if dest is not empty.

MathOptPresolve.presolve! — Method

presolve(pb::ProblemData)

Perform pre-solve.

MathOptPresolve.remove_empty_columns! — Method

remove_empty_columns!(ps::PresolveData)

Remove all empty columns.

Called once at the beginning of the presolve procedure. If an empty column is created later, it is removed on the spot.

MathOptPresolve.remove_forcing_rows! — Method

remove_forcing_rows!

Remove forcing and dominated row

MathOptPresolve.remove_free_column_singletons! — Method

remove_free_column_singletons!(ps)